added S3Cleanup feature

+++ src/main/java/org/cobbzilla/s3s3util/S3ToS3Master.java
@@ -47,7 +47,7 @@ public class S3ToS3Master {
         try {
             copyMaster.start();
 
-            if (context.getOptions().isDeleteRemoved()) {
+            if (!context.getOptions().isDelete() && context.getOptions().isDeleteRemoved()) {
                 deleteMaster = new DeleteMaster(client, context, workQueue, executorService);
                 deleteMaster.start();
             }
